Telegram_client/requirements.txt
pkg-resources==0.0.0
pyaes==1.6.1
pyasn1==0.4.8
PySocks==1.7.1
python-dotenv==0.17.0
rsa==4.7.2
socks==0
Telethon==1.21.1
